% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ahn}
\alias{ahn}
\title{ahn: Maaiveldhoogte in het onderzoeksgebied 'Rome' (Friesland).}
\format{
An object of class \code{RasterLayer} of dimension 771 x 386 x 1.
}
\usage{
ahn
}
\description{
Bron: AHN5 (resolutie 5x5 meter); (\url{https://www.ahn.nl/})
Eenheid: m+NAP.
}
\keyword{datasets}
