% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataBlauwgraslandRome.R
\docType{package}
\name{DataBlauwgraslandRome}
\alias{DataBlauwgraslandRome}
\title{Package "DataBlauwgraslandRome"}
\description{
Dit R-package bevat datasets die betrekking hebben op het gebied 'Rome' (Friesland).
}
\section{Datasets}{

Datasets:
\itemize{
\item \code{stk}
\item \code{ahn}
\item \code{sampled_area}
\item \code{perceel}
}

Meer informatie over deze datasets: ?stk, ?ahn, ?sampled_area of ?perceel
}

