% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{perceel}
\alias{perceel}
\title{perceel: Kaart met punten voor een mogelijke optimalisatie van de GxG.}
\format{
An object of class \code{RasterLayer} of dimension 771 x 386 x 1.
}
\usage{
perceel
}
\description{
1: mogelijke optimalisatie. Anders: NA.
Resolutie 5x5 meter.
}
\keyword{datasets}
