% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sampled_area}
\alias{sampled_area}
\title{sampled_area: Kaart van sample-punten in het gebied 'Rome' (Friesland).}
\format{
An object of class \code{RasterLayer} of dimension 771 x 386 x 1.
}
\usage{
sampled_area
}
\description{
In de sample punten (resolutie 5x5 meter) met de waarde 1 is van \emph{alle} rasters in de raster stack "stk" informatie aanwezig.
in de sample punten met een waarde 0 is (in een deel van de rasters of van alle rasters) deze informatie niet aanwezig.
}
\keyword{datasets}
