% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stk}
\alias{stk}
\title{stk}
\format{
An object of class \code{RasterStack} of dimension 771 x 386 x 9.
}
\usage{
stk
}
\description{
Kaarten met informatie over het gebied 'Rome' (Friesland).
}
\details{
\itemize{
  \item \code{Veg}: 1 = blauwgrasland; 0 = geen blauwgrasland.
  \item \code{GVG}, \code{GLG}, \code{GHG}: resp. gemiddelde voorjaarsgrondwaterstand, laagste grondwaterstand en hoogste grondwaterstand (cm-mv).
  \item \code{GVGtovKl}, \code{GLGtovKl}: resp. Gemiddelde voorjaarsgrondwaterstand en laagste grondwaterstand (cm tov keileemniveau).
  \item \code{Gt}: grondwatertrap: 10=Gt I; 20=Gt II; 25=Gt II-ster ; 30=Gt III; 35=Gt III-ster; 40=Gt IV; 45=Gt IV-ster; 50=Gt V;
    55=Gt V-ster; 60=Gt VI; 70=Gt VII; 75=Gt VII-ster
  \item \code{Bofek}: Bofek codes (\url{https://www.wur.nl/nl/show/Bodemkaart-1-50-000.htm})
  \item \code{DiepteKl}: Diepte van de keileem (cm=mv)
}
}
\examples{
\dontrun{plot(stk$Veg)}

}
\keyword{datasets}
